split code + started hub connection implementation
