+++ sample/Sample.Orleans.Client.WebApp/ClientApp/app/components/navmenu/navmenu.component.html
@@ -32,6 +32,11 @@
 					</a>
 				</li>
 				<li [routerLinkActive]="['link-active']">
+					<a [routerLink]="['/raw-sample']">
+						<span class="glyphicon glyphicon-education"></span> Raw Sample
+					</a>
+				</li>
+				<li [routerLinkActive]="['link-active']">
 					<a [routerLink]="['/fetch-data']">
 						<span class="glyphicon glyphicon-th-list"></span> Fetch data
 					</a>
